+++ src/System/Wlog/LogHandler/Roller.hs
@@ -11,12 +11,13 @@ import Universum
 
 import Control.Concurrent (modifyMVar, modifyMVar_, withMVar)
 import Data.Text.Lazy.Builder as B
-import Formatting (sformat, shown, (%))
+import Fmt ((+||), (||+))
 
 import System.Directory (removeFile, renameFile)
 import System.FilePath ((<.>))
 import System.IO (Handle, IOMode (ReadWriteMode), SeekMode (AbsoluteSeek, SeekFromEnd), hClose,
                   hFileSize, hFlush, hSeek)
+
 import System.Wlog.FileUtils (whenExist)
 import System.Wlog.Formatter (LogFormatter, nullFormatter)
 import System.Wlog.LoggerConfig (RotationParameters (..), isValidRotation)
@@ -49,7 +50,7 @@ instance LogHandler RollerHandler where
     emit rh bldr _    = liftIO $ rhWriteAction rh (error "Handler is used internally") (toText . B.toLazyText $ bldr)
     close RollerHandler{..} = liftIO $ withMVar rhFileHandle rhCloseAction
 
-data InvalidRotation = InvalidRotation !Text
+newtype InvalidRotation = InvalidRotation Text
     deriving (Show, Eq)
 
 instance Exception InvalidRotation
@@ -58,7 +59,7 @@ logIndex :: FilePath -> Int -> FilePath
 logIndex handlerPath i = handlerPath <.> show i
 
 rollerReadback :: MonadIO m => RollerHandler -> Int -> m [Text]
-rollerReadback RollerHandler{..} logsNum = liftIO $ do
+rollerReadback RollerHandler{..} logsNum = liftIO $
     modifyMVar rhFileHandle $ \h -> do
         hFlush h
         hSeek h AbsoluteSeek 0
@@ -90,7 +91,7 @@ rollerWriting RotationParameters{..} handlerPath loggingAction varHandle _ msg =
             for_ [lastIndex - 1,lastIndex - 2 .. 0] $ \i -> do
                 let oldLogFile = logIndex handlerPath i
                 let newLogFile = logIndex handlerPath (i + 1)
-                whenExist oldLogFile $ (`renameFile` newLogFile)
+                whenExist oldLogFile (`renameFile` newLogFile)
             let zeroIndex = logIndex handlerPath 0
             renameFile handlerPath zeroIndex
             let lastLogFile = logIndex handlerPath lastIndex
@@ -108,7 +109,7 @@ rotationFileHandler
     -> m RollerHandler
 rotationFileHandler rp@RotationParameters{..} _ _
     | not $ isValidRotation rp = liftIO $ throwM $ InvalidRotation $
-      sformat ("Rotation parameters must be positive: "%shown) rp
+      "Rotation parameters must be positive: "+||rp||+""
 rotationFileHandler rp@RotationParameters{..} handlerPath rhSeverities = liftIO $ do
     GenericHandler{..} <- fileHandler handlerPath rhSeverities
     rhFileHandle       <- newMVar privData