+++ src/System/Wlog/LoggerName.hs
@@ -2,23 +2,17 @@
 
 module System.Wlog.LoggerName
        ( LoggerName (..)
-       , loggerNameF
        ) where
 
-import           Universum
+import Universum
 
-import           Data.Hashable       (Hashable)
-import           Data.Semigroup      (Semigroup)
-import qualified Data.Semigroup      as Semigroup
-import           Data.String         (IsString)
-import           Data.Text.Buildable (Buildable)
-import qualified Data.Text.Buildable as Buildable
-import           Formatting          (Format, bprint, build, stext)
+import Data.Aeson.Types (ToJSON, ToJSONKey (..), toJSONKeyText)
+import Fmt (build)
 
 -- | Logger name to keep in context.
 newtype LoggerName = LoggerName
     { getLoggerName :: Text
-    } deriving (Show, IsString, Eq, Ord, Hashable)
+    } deriving (Show, IsString, Eq, Ord, Hashable, Generic)
 
 -- | Defined such that @n1@ is parent for @(n1 <> n2)@
 -- (see <http://hackage.haskell.org/package/hslogger-1.2.10/docs/System-Log-Logger.html hslogger description>).
@@ -30,11 +24,12 @@ instance Semigroup LoggerName where
 
 instance Monoid LoggerName where
     mempty = ""
-    mappend = (Semigroup.<>)
+    mappend = (<>)
 
 instance Buildable LoggerName where
-    build = bprint stext . getLoggerName
+    build = build . getLoggerName
 
--- | 'LoggerName' formatter which restricts type.
-loggerNameF :: Format r (LoggerName -> r)
-loggerNameF = build
+instance ToJSON LoggerName
+
+instance ToJSONKey LoggerName where
+    toJSONKey = toJSONKeyText getLoggerName